Relocate syn api settings
